requirements.txt
colorama==0.4.4
Django==2.2.23
django-barcode==0.1.1
django-bootstrap==0.2.4
django-bootstrap3==15.0.0
django-qr-code==2.1.0
importlib-metadata==2.1.1
Pillow==8.2.0
python-barcode==0.9.0
pytz==2021.1
qrcode==6.1
segno==1.3.3
six==1.16.0
sqlparse==0.4.1
zipp==3.4.1
